+++ packages/system/src/Effect/index.ts
@@ -133,7 +133,6 @@ export * from "./repeat"
 export * from "./repeatN"
 export * from "./repeatUntil"
 export * from "./repeatWhile"
-export * from "./result"
 export * from "./retry"
 export * from "./retryUntil"
 export * from "./retryWhile"
