+++ packages/system/src/Effect/some.ts
@@ -1,9 +1,8 @@
-import { pipe } from "../Function"
 import * as O from "../Option"
 import { succeed } from "./core"
 import type { Effect } from "./effect"
 import { fail } from "./fail"
-import { foldM } from "./foldM"
+import { foldM_ } from "./foldM_"
 
 /**
  * Converts an option on values into an option on errors.
@@ -11,11 +10,9 @@ import { foldM } from "./foldM"
 export function some<R, E, A>(
   self: Effect<R, E, O.Option<A>>
 ): Effect<R, O.Option<E>, A> {
-  return pipe(
+  return foldM_(
     self,
-    foldM(
-      (e) => fail(O.some(e)),
-      O.fold(() => fail(O.none), succeed)
-    )
+    (e) => fail(O.some(e)),
+    O.fold(() => fail(O.none), succeed)
   )
 }
