+++ packages/system/src/Effect/someOrElse.ts
@@ -1,14 +1,12 @@
-import { pipe } from "../Function"
 import * as O from "../Option"
 import type { Effect } from "./effect"
-import { map } from "./map"
+import { map_ } from "./map_"
 
 /**
  * Extracts the optional value, or returns the given 'orElse'.
  */
 export function someOrElse<B>(orElse: () => B) {
-  return <R, E, A>(self: Effect<R, E, O.Option<A>>) =>
-    pipe(self, map(O.getOrElse(orElse)))
+  return <R, E, A>(self: Effect<R, E, O.Option<A>>) => someOrElse_(self, orElse)
 }
 
 /**
@@ -18,5 +16,5 @@ export function someOrElse_<R, E, A, B>(
   self: Effect<R, E, O.Option<A>>,
   orElse: () => B
 ): Effect<R, E, A | B> {
-  return pipe(self, map(O.getOrElse(orElse)))
+  return map_(self, O.getOrElse(orElse))
 }
