+++ packages/system/src/Effect/someOrElseM.ts
@@ -1,6 +1,6 @@
-import { constant, flow, pipe } from "../Function"
+import { constant, flow } from "../Function"
 import * as O from "../Option"
-import { chain, succeed } from "./core"
+import { chain_, succeed } from "./core"
 import type { Effect } from "./effect"
 
 /**
@@ -17,8 +17,8 @@ export function someOrElseM_<R, E, A, R2, E2, B>(
   self: Effect<R, E, O.Option<A>>,
   orElse: Effect<R2, E2, B>
 ): Effect<R & R2, E | E2, A | B> {
-  return pipe(
+  return chain_(
     self as Effect<R, E, O.Option<A | B>>,
-    chain(flow(O.map(succeed), O.getOrElse(constant(orElse))))
+    flow(O.map(succeed), O.getOrElse(constant(orElse)))
   )
 }
