+++ packages/system/src/Effect/someOrFail.ts
@@ -1,6 +1,5 @@
-import { pipe } from "../Function"
 import * as O from "../Option"
-import { chain, chain_, effectTotal, succeed } from "./core"
+import { chain_, effectTotal, succeed } from "./core"
 import type { Effect } from "./effect"
 import { fail } from "./fail"
 
@@ -18,5 +17,8 @@ export function someOrFail_<R, E, A, E2>(
   self: Effect<R, E, O.Option<A>>,
   orFail: () => E2
 ): Effect<R, E | E2, A> {
-  return pipe(self, chain(O.fold(() => chain_(effectTotal(orFail), fail), succeed)))
+  return chain_(
+    self,
+    O.fold(() => chain_(effectTotal(orFail), fail), succeed)
+  )
 }
