+++ packages/system/src/Exit/_internal/effect.ts
@@ -1,3 +1,2 @@
+export { result, succeed } from "../../Effect/core"
 export { Effect } from "../../Effect/effect"
-export { succeed } from "../../Effect/core"
-export { result } from "../../Effect/result"
